backend/requirements/dev.txt
Django==5.2.9
python-decouple==3.8
django-cors-headers==4.9.0
django-rosetta==0.10.3
psycopg2-binary==2.9.11
pillow==12.1.1
redis==7.3.0
drf-spectacular==0.29.0
django-modeltranslation==0.20.2
httpx==0.28.1

django-debug-toolbar==5.0.1